requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.10.4
    # via -r requirements.in
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
asyncpg==0.27.0
    # via -r requirements.in
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via uvicorn
cryptography==40.0.2
    # via python-jose
dnspython==2.3.0
    # via email-validator
docker==6.1.2
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
email-validator==2.0.0.post2
    # via fastapi
fastapi[all]==0.95.1
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.0
    # via fastapi
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
itsdangerous==2.1.2
    # via fastapi
jinja2==3.1.2
    # via fastapi
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
orjson==3.8.12
    # via fastapi
packaging==23.1
    # via docker
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via fastapi
python-dotenv==1.0.0
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via
    #   -r requirements.in
    #   fastapi
pyyaml==6.0
    # via
    #   fastapi
    #   uvicorn
requests==2.30.0
    # via docker
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.13
    # via
    #   -r requirements.in
    #   alembic
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via
    #   alembic
    #   pydantic
    #   sqlalchemy
ujson==5.7.0
    # via fastapi
urllib3==2.0.2
    # via
    #   docker
    #   requests
uvicorn[standard]==0.22.0
    # via fastapi
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websocket-client==1.5.1
    # via docker
websockets==11.0.3
    # via uvicorn
